Fix panic in LibreQoS.py on update if an IP address mapping needed deleting.
* Adjust the Python integration `delete_ip_mapping` function to
  not require a secondary "upload" parameter - because the
  Python code is unaware of whether there needs to be a
  separation of the two at this point.
* Change ENOEXIST return code in BPF map delete to NOT be an
  error - it indicates that there was nothing to do, rather
  than something not working.

+++ src/rust/lqos_sys/src/bpf_map.rs
@@ -133,7 +133,12 @@ where
         let key_ptr: *mut K = key;
         let err = unsafe { bpf_map_delete_elem(self.fd, key_ptr as *mut c_void) };
         if err != 0 {
-            Err(Error::msg("Unable to delete from map"))
+            if err == -2 {
+                // ENOEXIST : not actually an error, just nothing to do
+                Ok(())
+            } else {
+                Err(Error::msg("Unable to delete from map"))
+            }
         } else {
             Ok(())
         }
